Properly trigger change detection when parent is not listening on resize
If you were to put this code in a parent component of lx-ellipsis,
all the ChangeDetectorRef.detectChanges() calls you see
being added in this commit would not be necessary:

@HostListener('window:resize', ['$event'])
onResize(event) {}

This is because the angular async pipe just calls the "markForCheck" method when an observable
it has subscribed to emits a new value.
This means the next time its next non-push changedetection
ancestor triggers change detection via zone.js, is the time when lx-ellipsis will do change detection as well.

In order for this component to also work in an environment where
no eventlistener for window.resize is active in a parent component,
we do the manual detectChanges() calls.

By adding a debounceTime on the "isContentOverFlowing$" observable we also save a lot of change detection calls
while users are "in the process" of resizing.